FSCORE-325
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@12533 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch.c
@@ -628,7 +628,7 @@ int main(int argc, char *argv[])
 			rlp.rlim_max = SWITCH_THREAD_STACKSIZE;
 			setrlimit(RLIMIT_STACK, &rlp);
 			fprintf(stderr, "Error: stacksize %ld is too large: run ulimit -s %d or run %s -waste.\nauto-adjusting stack size for optimal performance....\n", 
-					rlp.rlim_max / 1024, SWITCH_THREAD_STACKSIZE / 1024, argv[0]);
+					(int)(rlp.rlim_max / 1024), SWITCH_THREAD_STACKSIZE / 1024, argv[0]);
 			apr_terminate();
 			ret = (int)execv(argv[0], argv);
